+++ gameobject.py
@@ -1,0 +1,13 @@
+import pygame 
+pygame.init()
+
+class GameObject(pygame.sprite.Sprite):
+    def __init__(self, x, y, image):
+        super(GameObject, self).__init__()
+        self.surf = pygame.image.load(image)
+        self.x = x
+        self.y = y
+
+    def render(self, screen):
+        screen.blit(self.surf, (self.x, self.y))
+
